refactor(ft): refactor and format according to the norm

+++ include/minishell.h
@@ -151,7 +151,7 @@ void	freecmd(t_cmd *cmd);
 // Ft
 char  *ft_strchr(const char *s, int c);
 char **ft_split(char *str, char chr);
-int	   ft_dprintf(int fd, const char *format, ...);
+size_t ft_dprintf(int fd, const char *format, ...);
 size_t ft_strlen(const char *s);
 void  *ft_memset(void *s, int c, size_t n);
